Add slash command with message-ready output
